+++ netpath/web/service.py
@@ -465,7 +465,10 @@ class Service:
         )
         self.collector = Collector(self.flow_db, log=self.log)
         self.syslog = SyslogCollector(self.syslog_db, log=self.log)
-        self.snmp = TrapCollector(self.snmp_db, log=self.log, nodes_db=self.nodes_db)
+        # The lambda defers the node_poller lookup: it is built below.
+        self.snmp = TrapCollector(
+            self.snmp_db, log=self.log, nodes_db=self.nodes_db,
+            poll_now=lambda device_id: self.node_poller.poll_now(device_id))
         self.ipam = IpamWorker(self.ipam_db, log=self.log,
                                global_settings=lambda: self.settings,
                                nodes_db=self.nodes_db)